ctfs/ImaginaryCTF/2023/web/Perfect_Picture/requirements.txt
flask==2.3.2
pillow==10.0.0
pyexiftool==0.5.5
